requirements.txt
pyTelegramBotAPI
python-dotenv
pymongo
beautifulsoup4
langchain
langchain-community
langchain-core
langchain-anthropic
langchain-openai
chromadb
colorama
pymupdf
pytesseract
pdf2image
pydub
reportlab
pillow
pdfminer.six